requirements.txt
sqlalchemy
psycopg2-binary
pyyaml
pandas
